+++ app/src/main/java/com/tqeil/couplesenppai/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.tqeil.couplesenppai
+package com.tqeil.couplesenppai.ui.main
 
 import android.os.Bundle
 import android.support.design.widget.Snackbar
@@ -9,6 +9,7 @@ import android.support.v7.app.ActionBarDrawerToggle
 import android.support.v7.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
+import com.tqeil.couplesenppai.R
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.app_bar_main.*
 import java.lang.IllegalArgumentException
@@ -29,7 +30,9 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         }
 
         val toggle = ActionBarDrawerToggle(
-            this, drawer_layout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close
+            this, drawer_layout, toolbar,
+            R.string.navigation_drawer_open,
+            R.string.navigation_drawer_close
         )
         drawer_layout.addDrawerListener(toggle)
         toggle.syncState()
