+++ app/src/main/java/com/tqeil/couplesenppai/ui/main/fragment/FeedFragment.kt
@@ -1,17 +1,18 @@
-package com.tqeil.couplesenppai
+package com.tqeil.couplesenppai.ui.main.fragment
 
 import android.support.v4.app.Fragment
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.tqeil.couplesenppai.R
 
-class CalendarFragment: Fragment() {
+class FeedFragment: Fragment() {
 
     companion object {
 
-        fun create(): CalendarFragment {
-            val fragment = CalendarFragment()
+        fun create(): FeedFragment {
+            val fragment = FeedFragment()
             val args = Bundle()
             fragment.arguments = args
             return fragment
@@ -20,7 +21,7 @@ class CalendarFragment: Fragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val rootView = inflater.inflate(R.layout.fragment_calendar, container, false)
+        val rootView = inflater.inflate(R.layout.fragment_feed, container, false)
 
         return rootView
     }
