+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".ui.main.MainActivity">
 
     <android.support.design.widget.AppBarLayout
             android:layout_height="wrap_content"
