+++ app/src/main/res/layout/content_main.xml
@@ -7,7 +7,7 @@
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
         tools:showIn="@layout/app_bar_main"
-        tools:context=".MainActivity">
+        tools:context=".ui.main.MainActivity">
 
     <android.support.v4.view.ViewPager
             android:id="@+id/pager"
